remove goonjcustomeventcron
